Review fix: Renamed and Replaced components.
