--- src/components/ApplicationFooter/index.js
@@ -1,3 +1,0 @@
-import ApplicationFooter from './ApplicationFooter';
-
-export default ApplicationFooter;
