--- src/components/ApplicationHeader/index.js
@@ -1,3 +1,0 @@
-import ApplicationHeader from './ApplicationHeader';
-
-export default ApplicationHeader;
