--- src/components/ApplicationLayout/index.js
@@ -1,3 +1,0 @@
-import ApplicationLayout from './ApplicationLayout';
-
-export default ApplicationLayout;
