+++ src/pages/404.js
@@ -1,16 +1,16 @@
 import React from "react";
 
-import ApplicationLayout from "@components/ApplicationLayout";
+import Layout from "@components/Layout";
 import Head from "@components/Head";
 
 const NotFound = () => (
-  <ApplicationLayout>
+  <Layout>
     <Head>
       <title>Not Found - Halo Lab Blog</title>
     </Head>
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </ApplicationLayout>
+  </Layout>
 );
 
 export default NotFound;
