+++ src/pages/thanks.js
@@ -1,17 +1,17 @@
 import React from 'react';
 
-import ApplicationLayout from "@components/ApplicationLayout";
+import Layout from "@components/Layout";
 import Head from "@components/Head";
 
 const Thanks = () => {
 
   return (
-    <ApplicationLayout>
+    <Layout>
       <Head>
         <title>Thanks - Halo Lab Blog</title>
       </Head>
       <div>Thanks</div>
-    </ApplicationLayout>
+    </Layout>
   );
 };
